more generics
git-svn-id: https://svn.apache.org/repos/asf/maven/doxia/doxia/trunk@1091053 13f79535-47bb-0310-9956-ffa450edef68

+++ doxia-modules/doxia-module-itext/src/main/java/org/apache/maven/doxia/module/itext/SinkActionContext.java
@@ -132,7 +132,7 @@ public class SinkActionContext
         }
         else
         {
-            return ( (Integer) stack.peek() ).intValue();
+            return stack.peek().intValue();
         }
     }
